commit antes de crear rama

+++ spring-boot-hoyos/src/main/java/com/cuevasdeayllon/controllers/AnunciosController.java
@@ -1,6 +1,5 @@
 package com.cuevasdeayllon.controllers;
 
-
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -41,71 +40,73 @@ public class AnunciosController {
 	private DocumentosRepositoryImpl documentosRepositoryImpl;
 
 	@GetMapping("/todosAnuncios")
-	public String listaDeAnuncios( Model model) {
+	public String listaDeAnuncios(Model model) {
 
 		logger.info("Entramos en metodo /todosAnuncios");
 
 		try {
-			List <Anuncios>anuncios= anuncioRepositoryImpl.listAnuncio();
+			List<Anuncios> anuncios = anuncioRepositoryImpl.listAnuncio();
 
-			anuncios=anuncios.stream().sorted(Comparator.comparing(Anuncios::getIdAnuncios)
-					.reversed()).collect(Collectors.toList());
+			anuncios = anuncios.stream().sorted(Comparator.comparing(Anuncios::getIdAnuncios).reversed())
+					.collect(Collectors.toList());
 
-			model.addAttribute("listaAnuncios",anuncios);
+			model.addAttribute("listaAnuncios", anuncios);
 
 			return "tablonanuncios";
 
-		}catch(Exception e) {
+		} catch (Exception e) {
 
-			logger.info("El error que da en /todosAnuncios es: "+e.getMessage());
+			logger.info("El error que da en /todosAnuncios es: " + e.getMessage());
 
 		}
 		return "tablonanuncios";
 
 	}
-	@PostMapping("/subirAnuncio")
-	public String insertarAnuncio( @RequestParam("titulo")String titulo,@RequestParam("anuncio")String anuncio,@RequestParam("file")MultipartFile foto, Model model) {
 
-		//String rootPath="/uploadsAnuncios/";
-		String rootPath="C://TEMP//uploadsAnuncios";
+	@PostMapping("/subirAnuncio")
+	public String insertarAnuncio(@RequestParam("titulo") String titulo, @RequestParam("anuncio") String anuncio,
+			@RequestParam("file") MultipartFile foto, Model model) {
+
+		// String rootPath="/uploadsAnuncios/";
+		String rootPath = "C://TEMP//uploadsAnuncios";
 		logger.info("Entramos en metodo /subirAnuncio");
 
 		int oraLen = foto.getOriginalFilename().length();
-		logger.info("El nombre de la foto es: "+foto.getOriginalFilename());
+		logger.info("El nombre de la foto es: " + foto.getOriginalFilename());
 
-		for (int i = 0; i <  oraLen; i++) {
+		for (int i = 0; i < oraLen; i++) {
 			if (foto.getOriginalFilename().charAt(i) == ' ') {
-				model.addAttribute("faltaTitulo", "El nombre de la foto no puede tener espacios en blanco.");
+				model.addAttribute("fotoConEspacio", "El nombre de la foto no puede tener espacios en blanco.");
 				return "subirAnuncio";
 
 			}
 		}
 
-
 		try {
-			if(titulo.isEmpty()) {
+			if (titulo.isEmpty()) {
 
-				model.addAttribute("faltaTitulo", "El titulo en los anuncios es necesario, lo unico que se puede omitir es la foto.");
+				model.addAttribute("faltaTitulo",
+						"El titulo en los anuncios es necesario, lo unico que se puede omitir es la foto.");
 				return "subirAnuncio";
 			}
 
-			else if(anuncio.isEmpty()) {
+			else if (anuncio.isEmpty()) {
 
-				model.addAttribute("faltaAnuncio", "El texto en los anuncios es necesario, lo unico que se puede omitir es la foto.");
+				model.addAttribute("faltaAnuncio",
+						"El texto en los anuncios es necesario, lo unico que se puede omitir es la foto.");
 				return "subirAnuncio";
-			}else {
-				Anuncios anuncioeditable=new Anuncios();
+			} else {
+				Anuncios anuncioeditable = new Anuncios();
 				anuncioeditable.setIdAnuncios(0);
 				anuncioeditable.setAnuncio(anuncio);
-				anuncioeditable.setFecha(new Date());		
+				anuncioeditable.setFecha(new Date());
 				anuncioeditable.setTitulo_anuncio(titulo);
-				if(!foto.isEmpty()) {
-
+				if (!foto.isEmpty()) {
 
 					try {
-						byte[]bytes=foto.getBytes();
-						Path rutaCompleta=Paths.get(rootPath+"//"+foto.getOriginalFilename());
-						Files.write(rutaCompleta,bytes);
+						byte[] bytes = foto.getBytes();
+						Path rutaCompleta = Paths.get(rootPath + "//" + foto.getOriginalFilename());
+						Files.write(rutaCompleta, bytes);
 
 						anuncioeditable.setFoto_anuncio(foto.getOriginalFilename());
 					} catch (IOException e) {
@@ -115,83 +116,83 @@ public class AnunciosController {
 
 				}
 
-
-				anuncioRepositoryImpl.insertarAnucio(anuncioeditable,foto);
+				anuncioRepositoryImpl.insertarAnucio(anuncioeditable, foto);
 
 				model.addAttribute("anuncioSubido", "El anuncio se ha agregado con exito.");
 
 				return "subirAnuncio";
 			}
-		}catch(Exception e) {
+		} catch (Exception e) {
 
 			model.addAttribute("anuncioSubido", "El anuncio no se ha agregado.");
 
-			logger.info("El error que da en /subirAnuncio es: "+e.getMessage());
+			logger.info("El error que da en /subirAnuncio es: " + e.getMessage());
 		}
 		return "subirAnuncio";
 
 	}
+
 	@PostMapping("/editarAnuncio")
-	public String editarAnuncio( @RequestParam("idAnuncio")int idAnuncio, @RequestParam("titulo")String titulo,@RequestParam("fecha")String fecha,@RequestParam("anuncio")String anuncio,@RequestParam("file")MultipartFile foto, Model model) {
+	public String editarAnuncio(@RequestParam("idAnuncio") int idAnuncio, @RequestParam("titulo") String titulo,
+			@RequestParam("fecha") String fecha, @RequestParam("anuncio") String anuncio,
+			@RequestParam("file") MultipartFile foto, Model model) {
 
-		//String rootPath="/uploadsAnuncios/";
-		String rootPath="C://TEMP//uploadsAnuncios";
+		// String rootPath="/uploadsAnuncios/";
+		String rootPath = "C://TEMP//uploadsAnuncios";
 		logger.info("Entramos en metodo /subirAnuncio");
-		Anuncios anuncioeditable=anuncioRepositoryImpl.recuperarAnuncio(idAnuncio);
+		Anuncios anuncioeditable = anuncioRepositoryImpl.recuperarAnuncio(idAnuncio);
 		int oraLen = foto.getOriginalFilename().length();
-		logger.info("El nombre de la foto es: "+foto.getOriginalFilename());
+		logger.info("El nombre de la foto es: " + foto.getOriginalFilename());
 
-		for (int i = 0; i <  oraLen; i++) {
+		for (int i = 0; i < oraLen; i++) {
 			if (foto.getOriginalFilename().charAt(i) == ' ') {
 				model.addAttribute("faltaTitulo", "El nombre de la foto no puede tener espacios en blanco.");
-				List <Anuncios>anuncios=anuncioRepositoryImpl.listAnuncio();
+				List<Anuncios> anuncios = anuncioRepositoryImpl.listAnuncio();
 
-
-				model.addAttribute("listaAnuncios",anuncios);
+				model.addAttribute("listaAnuncios", anuncios);
 				return "listaAnuncios";
 
 			}
 		}
 		try {
-			if(titulo.isEmpty()) {
+			if (titulo.isEmpty()) {
 
-				model.addAttribute("faltaTitulo", "El titulo en los anuncios es necesario, lo unico que se puede omitir es la foto.");
-				List <Anuncios>anuncios=anuncioRepositoryImpl.listAnuncio();
+				model.addAttribute("faltaTitulo",
+						"El titulo en los anuncios es necesario, lo unico que se puede omitir es la foto.");
+				List<Anuncios> anuncios = anuncioRepositoryImpl.listAnuncio();
 
-
-				model.addAttribute("listaAnuncios",anuncios);
+				model.addAttribute("listaAnuncios", anuncios);
 				return "listaAnuncios";
 			}
 
-			else if(anuncio.isEmpty()) {
-				List <Anuncios>anuncios=anuncioRepositoryImpl.listAnuncio();
+			else if (anuncio.isEmpty()) {
+				List<Anuncios> anuncios = anuncioRepositoryImpl.listAnuncio();
 
+				model.addAttribute("listaAnuncios", anuncios);
 
-				model.addAttribute("listaAnuncios",anuncios);
-
-				model.addAttribute("faltaAnuncio", "El texto en los anuncios es necesario, lo unico que se puede omitir es la foto.");
+				model.addAttribute("faltaAnuncio",
+						"El texto en los anuncios es necesario, lo unico que se puede omitir es la foto.");
 				return "listaAnuncios";
-			}else {
+			} else {
 				SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-				java.sql.Date fechaConvertida=null;
+				java.sql.Date fechaConvertida = null;
 				try {
-				    Date parsed =  dateFormat.parse(fecha);
-				    fechaConvertida = new java.sql.Date(parsed.getTime());
-				} catch(Exception e) {
-				    System.out.println("Error occurred"+ e.getMessage());
+					Date parsed = dateFormat.parse(fecha);
+					fechaConvertida = new java.sql.Date(parsed.getTime());
+				} catch (Exception e) {
+					System.out.println("Error occurred" + e.getMessage());
 				}
-				
+
 				anuncioeditable.setIdAnuncios(idAnuncio);
 				anuncioeditable.setAnuncio(anuncio);
-				anuncioeditable.setFecha( fechaConvertida);		
+				anuncioeditable.setFecha(fechaConvertida);
 				anuncioeditable.setTitulo_anuncio(titulo);
-				if(!foto.isEmpty()) {
-
+				if (!foto.isEmpty()) {
 
 					try {
-						byte[]bytes=foto.getBytes();
-						Path rutaCompleta=Paths.get(rootPath+"//"+foto.getOriginalFilename());
-						Files.write(rutaCompleta,bytes);
+						byte[] bytes = foto.getBytes();
+						Path rutaCompleta = Paths.get(rootPath + "//" + foto.getOriginalFilename());
+						Files.write(rutaCompleta, bytes);
 
 						anuncioeditable.setFoto_anuncio(foto.getOriginalFilename());
 					} catch (IOException e) {
@@ -202,118 +203,107 @@ public class AnunciosController {
 				}
 
 				anuncioRepositoryImpl.editarAnuncio(anuncioeditable.getIdAnuncios(), anuncioeditable);
-				
-				List <Anuncios>anuncios=anuncioRepositoryImpl.listAnuncio();
 
+				List<Anuncios> anuncios = anuncioRepositoryImpl.listAnuncio();
 
-				model.addAttribute("listaAnuncios",anuncios);
+				model.addAttribute("listaAnuncios", anuncios);
 
 				model.addAttribute("anuncioSubido", "El anuncio se ha editado con exito.");
 
 				return "listaAnuncios";
 			}
-		}catch(Exception e) {
+		} catch (Exception e) {
 
 			model.addAttribute("anuncioSubido", "El anuncio no se ha agregado.");
 
-			logger.info("El error que da en /subirAnuncio es: "+e.getMessage());
+			logger.info("El error que da en /subirAnuncio es: " + e.getMessage());
 		}
-		List <Anuncios>anuncios=anuncioRepositoryImpl.listAnuncio();
+		List<Anuncios> anuncios = anuncioRepositoryImpl.listAnuncio();
 
-
-		model.addAttribute("listaAnuncios",anuncios);
+		model.addAttribute("listaAnuncios", anuncios);
 		return "listaAnuncios";
 
 	}
+
 	@GetMapping("/listaAnunciosAdmin")
-	public String listaDeAnunciosAdministrador( Model model) {
+	public String listaDeAnunciosAdministrador(Model model) {
 
 		logger.info("Entramos en metodo /listaAnunciosAdmin");
 		try {
-			List <Anuncios>anuncios=anuncioRepositoryImpl.listAnuncio();
+			List<Anuncios> anuncios = anuncioRepositoryImpl.listAnuncio();
 
-
-			model.addAttribute("listaAnuncios",anuncios);
-
+			model.addAttribute("listaAnuncios", anuncios);
 
 			return "listaAnuncios";
-		}catch(Exception e) {
-			logger.info("El error que da en /listaAnunciosAdmin es: "+e.getMessage());
+		} catch (Exception e) {
+			logger.info("El error que da en /listaAnunciosAdmin es: " + e.getMessage());
 		}
 		return "listaAnuncios";
 
 	}
+
 	@PostMapping("/borrarAnuncio")
-	public String borrarAnuncio(@RequestParam("idAnuncio")int idAnuncio, Model model) {
+	public String borrarAnuncio(@RequestParam("idAnuncio") int idAnuncio, Model model) {
 		try {
-			if(idAnuncio>0) {
+			Anuncios anuncios = anuncioRepositoryImpl.recuperarAnuncio(idAnuncio);
+			
+			if (idAnuncio > 0 && anuncios != null) {
 				anuncioRepositoryImpl.deleteAnuncio(idAnuncio);
+			} else {
+				return "listaAnuncios";
 			}
 
-		}catch(Exception e) {
+		} catch (Exception e) {
 
-			logger.info("El error que da en /borrarAnuncio es: "+e.getMessage());
+			logger.info("El error que da en /borrarAnuncio es: " + e.getMessage());
 
 		}
 
-		List <Anuncios>anuncios;
+		List<Anuncios> anuncios;
 		try {
-			anuncios=anuncioRepositoryImpl.listAnuncio();
-			model.addAttribute("listaAnuncios",anuncios);
-		}catch(Exception e) {
+			anuncios = anuncioRepositoryImpl.listAnuncio();
+			model.addAttribute("listaAnuncios", anuncios);
+		} catch (Exception e) {
 
-			logger.info("El error que da en /borrarAnuncio recuperar todos es: "+e.getMessage());
+			logger.info("El error que da en /borrarAnuncio recuperar todos es: " + e.getMessage());
 
 		}
 
 		return "listaAnuncios";
 
 	}
-	@PostMapping("/subirDocumento")
-	public String insertarDocumento(@RequestParam("titulo")String titulo, @RequestParam("file")MultipartFile foto, Model model) throws IOException {
 
+	@PostMapping("/subirDocumento")
+	public String insertarDocumento(@RequestParam("titulo") String titulo, @RequestParam("file") MultipartFile foto,
+			Model model) throws IOException {
 
 		logger.info("Entramos en metodo /Documento");
 
-
-		logger.info("El nombre de la foto es: "+foto.getOriginalFilename());
-
-
-
-
+		logger.info("El nombre de la foto es: " + foto.getOriginalFilename());
 
 		documentosRepositoryImpl.insertarDocumentos(titulo, foto);
 
 		model.addAttribute("anuncioSubido", "El anuncio se ha agregado con exito.");
 
-
-
-
 		return "fileUpload";
 
 	}
+
 	@GetMapping("/todosDocumentos")
 	public String todosDocumento(Model model) throws IOException {
 
-
 		logger.info("Entramos en metodo /Documento");
 
+		List<Documentos> documento = new ArrayList<>();
 
-		List<Documentos> documento=new ArrayList<>();
-
-
-		documento=documentosRepositoryImpl.listDocumentos();
-		documento=documento.stream().sorted(Comparator.comparing(Documentos::getId)
-				.reversed()).collect(Collectors.toList());
+		documento = documentosRepositoryImpl.listDocumentos();
+		documento = documento.stream().sorted(Comparator.comparing(Documentos::getId).reversed())
+				.collect(Collectors.toList());
 
 		model.addAttribute("listaDocumentos", documento);
-
-
-
 
 		return "documentos";
 
 	}
-
 
 }
